Start and stop debug session
